+++ server/src/test/typeIssues.test.ts
@@ -1166,7 +1166,10 @@ describe('Type Issues', () => {
 				await context.parser.stable;
 				const runtime = await context.getRuntime();
 				const issues = runtime.typesIssues;
-				expect(issues.length).toEqual(0);
+				expect(issues.length).toEqual(1);
+				expect(issues[0].raw.issues).toEqual([
+					StandardTypeIssue.EXPECTED_NODE_ADDRESS,
+				]);
 			});
 
 			test('valid type single string - node', async () => {